refactor(systems): relocate benchmark system registry to top-level q2mm/systems.py
De-misfiles the benchmark system registry (SYSTEMS, SystemData, SystemSpec,
load_system, plus loader helpers) out of q2mm/diagnostics/. It is core domain
configuration that benchmark_runner.py and the workflows/ package depend on,
not a diagnostic/reporting utility.

- git mv q2mm/diagnostics/systems.py -> q2mm/systems.py (history preserved)
- repoint every importer (source, tests, scripts, docstrings) from
  q2mm.diagnostics.systems -> q2mm.systems; grep oracle returns zero hits
- drop the four systems names from q2mm.diagnostics.__init__ (nothing imported
  them from the package top level; no re-export shim per AGENTS.md alpha rules)
- rename the mypy grandfather entry q2mm.diagnostics.systems -> q2mm.systems in
  pyproject.toml (same count, sorted position); pre-existing type errors stay
  grandfathered under the new name
- fix the two __file__-relative path anchors (_REPO_ROOT and the supporting-info
  dir) from parent.parent.parent to parent.parent so resolved paths stay
  byte-identical after moving up one directory level (zero behavior change)
- update docs/how-it-works/architecture.md module tree and the test guard, and
  repoint the stale q2mm/diagnostics/systems.py path references in AGENTS.md and
  docs/benchmarks/qfuerza-recovery.md to the new location

Zero-behavior-change move + import-path repoint only.
