+++ test/benchmarks/test_optimization.py
@@ -19,7 +19,7 @@ pytestmark = [pytest.mark.benchmark, pytest.mark.nightly]
 def _run_and_validate(engine: object) -> None:
     """Run the benchmark pipeline and validate structure."""
     from q2mm.diagnostics.benchmark import run_combo
-    from q2mm.diagnostics.systems import load_system
+    from q2mm.systems import load_system
 
     sys_data = load_system("ch3f", engine=engine)
     result = run_combo(
