+++ test/integration/test_backend_optimizer_matrix.py
@@ -186,7 +186,7 @@ class TestBenchmarkPipeline:
     def result(self) -> BenchmarkResult:
         from q2mm.backends.mm.openmm import OpenMMEngine
         from q2mm.diagnostics.benchmark import run_combo
-        from q2mm.diagnostics.systems import load_system
+        from q2mm.systems import load_system
 
         engine = OpenMMEngine()
         sys_data = load_system("ch3f", engine=engine)
